stash work

+++ src/module/instruction.rs
@@ -7,7 +7,7 @@ use crate::{utils, Error, Value};
 #[derive(Clone, PartialEq)]
 pub enum Instruction {
     Const(Value),
-    Call(u32),
+    Call(usize),
 }
 
 impl Instruction {
@@ -18,7 +18,7 @@ impl Instruction {
             0x41 => Ok(Some(Instruction::Const(Value::Integer32(
                 utils::read_leb128_i32(reader)?,
             )))),
-            0x10 => Ok(Some(Instruction::Call(utils::read_leb128_u32(reader)?))),
+            0x10 => Ok(Some(Instruction::Call(utils::read_leb128_u32(reader)? as usize))),
             x => panic!("Instruction not implemented: 0x{:X}", x),
         }
     }
